requirements.txt
# Core
django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3

# Security
bcrypt>=4.0

# AI / Vector Search
faiss-cpu>=1.7
numpy>=1.24
sentence-transformers>=3.0

# Document Processing
pypdf>=3.0

# HTTP
requests>=2.31
# Keep urllib3 v1 on Python environments linked against LibreSSL
# (avoids NotOpenSSLWarning seen on macOS system Python 3.9).
urllib3<2

# Environment
python-dotenv>=1.0

# Cloud Storage
cloudinary>=1.41
django-cloudinary-storage>=0.3

# CORS
django-cors-headers>=4.4
